Fixed sumthin ig, made ready event
